refactor: Remove `src/maintenance`
